requirements.txt
fastapi==0.111.0
numpy==1.23.5
onnxruntime==1.18.1
opencv_python==4.4.0.40
opencv_python_headless==4.2.0.34
Pillow==9.4.0
psutil==6.0.0
sahi==0.11.18
torch==2.3.1
ultralytics==8.2.56
uvicorn==0.30.1
ffmpeg-python==0.2.0
